Moved visit of Pattern up
